+++ src/decompositions/decompositions.cpp
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020-2021 INRIA
+ * Copyright 2020-2024 INRIA
  */
 
 #include "eigenpy/decompositions/decompositions.hpp"
@@ -7,6 +7,9 @@
 #include "eigenpy/decompositions/EigenSolver.hpp"
 #include "eigenpy/decompositions/LDLT.hpp"
 #include "eigenpy/decompositions/LLT.hpp"
+#include "eigenpy/decompositions/PermutationMatrix.hpp"
+#include "eigenpy/decompositions/sparse/LLT.hpp"
+#include "eigenpy/decompositions/sparse/LDLT.hpp"
 #include "eigenpy/decompositions/SelfAdjointEigenSolver.hpp"
 #include "eigenpy/decompositions/minres.hpp"
 #include "eigenpy/fwd.hpp"
@@ -14,6 +17,9 @@
 namespace eigenpy {
 void exposeDecompositions() {
   using namespace Eigen;
+
+  typedef Eigen::SparseMatrix<double, Eigen::ColMajor> ColMajorSparseMatrix;
+  //  typedef Eigen::SparseMatrix<double,Eigen::RowMajor> RowMajorSparseMatrix;
 
   EigenSolverVisitor<MatrixXd>::expose("EigenSolver");
   SelfAdjointEigenSolverVisitor<MatrixXd>::expose("SelfAdjointEigenSolver");
@@ -34,5 +40,21 @@ void exposeDecompositions() {
         .value("ABx_lx", ABx_lx)
         .value("BAx_lx", BAx_lx);
   }
+
+  // Expose sparse decompositions
+  {
+    SimplicialLLTVisitor<ColMajorSparseMatrix>::expose("SimplicialLLT");
+    SimplicialLDLTVisitor<ColMajorSparseMatrix>::expose("SimplicialLDLT");
+  }
+
+  PermutationMatrixVisitor<Eigen::Dynamic>::expose("PermutationMatrix");
+
+#ifdef EIGENPY_WITH_CHOLMOD_SUPPORT
+  exposeCholmod();
+#endif
+
+#ifdef EIGENPY_WITH_ACCELERATE_SUPPORT
+  exposeAccelerate();
+#endif
 }
 }  // namespace eigenpy